Create PUT - Zoho to Acuity
When the admin reschedules the appointment date and time without any restrictions, it is important for the admin to be careful when editing the appointment. Currently, only the date and time of the appointment can be edited, and this change will be reflected in the Acuity data.

I will also address any discrepancies in date and time due to time zones.